requirements.txt
torch>=2.2.0
torchvision>=0.17.0
albumentations>=1.4.0
opencv-python>=4.9.0.80
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
PyYAML>=6.0.1
tqdm>=4.66.0
Pillow>=10.0.0
PyPDF2>=3.0.0
